pyproject.toml
[tool.poetry]
name = "django-reviews"
version = "0.1.0"
description = ""
authors = ["nikkapral"]
license = "APACHE MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-rest-framework = "^0.1.0"
psycopg = "^3.2.3"
psycopg2 = "^2.9.10"
django-environ = "^0.11.2"
django-stubs = "^5.1.1"
django-ninja = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
